// project-Intership/src/app/shared/model/masterdata.ts
export class Masterdatas {
    groupName: string;
    childs: {
        key: string;
        value: string;
        displayText: string;
        code: string;
        typeGroup: string
    }[];
}
export class Masterdata {
    key: string;
    value: string;
    displayText: string;
    code: string;
    typeGroup: string;
}